20240702/requirements.txt
psycopg2-binary
requests
pydantic
python-dotenv